+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 
 # CHANGELOG
 
+# v3.0.0
+
+* [BREAKING] Renamed `Oauth2ApiClient::HttpError` to
+  `Oauth2ApiClient::ResponseError`
+* Added http error exception classes like e.g.
+  `Oauth2ApiClient::ResponseError::NotFound`,
+  `Oauth2ApiClient::ResponseError::InternalServerError`, etc.
+
 # v2.1.0
 
 * Include the response code and body in `HttpError#to_s`